pyproject.toml
[project]
name = "rag"
version = "0.6.3"
description = "RAG (Retrieval-Augmented Generation)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "boto3>=1.39.11",
  "dotenv>=0.9.9",
  "firecrawl-py>=2.16.1",
  "google-generativeai>=0.8.5",
  "humanize>=4.12.3",
  "langchain-community>=0.3.27",
  "langchain-text-splitters>=0.3.8",
  "markdownify>=1.1.0",
  "nltk>=3.9.1",
  "ollama>=0.5.1",
  "openai>=1.97.0",
  "rich>=14.0.0",
  "textual>=0.55.0",
  "tiktoken>=0.7.0",
  "unstructured[md]>=0.18.9",
  "colorlog>=6.8.2",
  "chromadb>=0.5.23",
  "flask>=3.1.0",
  "flask-socketio>=5.4.2",
  "flask-cors>=5.0.0",
  "validators>=0.35.0",
  "ruamel-yaml>=0.18.14",
  "keybert>=0.9.0",
  "trafilatura>=2.0.0",
]
